stock quote model

+++ Models/StockQuotes.cs
@@ -33,27 +33,27 @@ public class StockQuotes
     [BsonElement("price_change")]
     [JsonPropertyName("price_change")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? price_change { get; set; }
+    public Decimal? price_change { get; set; }
 
     [BsonElement("price_change_percent")]
     [JsonPropertyName("price_change_percent")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? price_change_percent { get; set; }
+    public Decimal? price_change_percent { get; set; }
 
     [BsonElement("market_day_high")]
     [JsonPropertyName("market_day_high")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? market_day_high { get; set; }
+    public Decimal? market_day_high { get; set; }
 
     [BsonElement("market_day_low")]
     [JsonPropertyName("market_day_low")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? market_day_low { get; set; }
+    public Decimal? market_day_low { get; set; }
 
     [BsonElement("market_open_price")]
     [JsonPropertyName("maket_open_price")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? market_open_price { get; set; }
+    public Decimal? market_open_price { get; set; }
 
     [BsonElement("latest_info_date")]
     [JsonPropertyName("latest_info_date")]
@@ -62,12 +62,12 @@ public class StockQuotes
     [BsonElement("previous_close_price")]
     [JsonPropertyName("previous_close_price")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? previous_close_price { get; set; }
+    public Decimal? previous_close_price { get; set; }
 
     [BsonElement("current_price")]
     [JsonPropertyName("current_price")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? current_price { get; set; }
+    public Decimal? current_price { get; set; }
 
     [BsonElement("volume")]
     [JsonPropertyName("volume")]
@@ -84,32 +84,32 @@ public class StockQuotes
     [BsonElement("bid")]
     [JsonPropertyName("bid")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? bid { get; set; }
+    public Decimal? bid { get; set; }
 
     [BsonElement("ask")]
     [JsonPropertyName("ask")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? ask { get; set; }
+    public Decimal? ask { get; set; }
 
     [BsonElement("ytd_change")]
     [JsonPropertyName("ytd_change")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? ytd_change { get; set; }
+    public Decimal? ytd_change { get; set; }
 
     [BsonElement("ytd_change_percent")]
     [JsonPropertyName("ytd_change_percent")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? ytd_change_percent { get; set; }
+    public Decimal? ytd_change_percent { get; set; }
 
     [BsonElement("fifty_two_week_high")]
     [JsonPropertyName("fifty_two_week_high")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? fifty_two_week_high { get; set; }
+    public Decimal? fifty_two_week_high { get; set; }
 
     [BsonElement("fifty_two_week_low")]
     [JsonPropertyName("fifty_two_week_low")]
     [BsonRepresentation(BsonType.Int32, AllowTruncation = true)]
-    public int? fifty_two_week_low { get; set; }
+    public Decimal? fifty_two_week_low { get; set; }
 
   
 }
